# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/domains": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all domains belonging to the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "Get user domains",
                "responses": {
                    "200": {
                        "description": "List of user domains",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domains.DomainResult"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Register a new domain for the authenticated user with user-provided encryption public key and optional webhook configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "Create a new domain",
                "parameters": [
                    {
                        "description": "Domain creation details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domains.CreateDomainRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Domain created successfully",
                        "schema": {
                            "$ref": "#/definitions/domains.DomainResult"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "402": {
                        "description": "Plan limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/domains.planLimitExceededResponse"
                        }
                    }
                }
            }
        },
...
        "/domains/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a specific domain by its ID (must belong to authenticated user)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "Get domain by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Domain ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Domain details",
                        "schema": {
                            "$ref": "#/definitions/domains.DomainResult"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid domain ID",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "403": {
                        "description": "Forbidden - domain does not belong to user",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Domain not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update domain settings including public key, verification status, and webhook configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "Update domain",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Domain ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Domain update details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domains.UpdateDomainRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Domain updated successfully",
                        "schema": {
                            "$ref": "#/definitions/domains.DomainResult"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "403": {
                        "description": "Forbidden - domain does not belong to user",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Domain not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a domain and all associated email addresses and received emails",
                "tags": [
                    "Domains"
                ],
                "summary": "Delete domain",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Domain ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Domain deleted successfully"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Domain not found or does not belong to user",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domains.CreateDomainRequest": {
            "type": "object",
            "required": [
                "domain",
                "public_key"
            ],
            "properties": {
                "auto_create_address": {
                    "type": "boolean"
                },
                "domain": {
                    "type": "string",
                    "maxLength": 253,
                    "minLength": 1
                },
                "public_key": {
                    "type": "string",
                    "minLength": 100
                },
                "storage_enabled": {
                    "type": "boolean"
                },
                "webhook_config": {
                    "$ref": "#/definitions/domains.WebhookConfigRequest"
                }
            }
        },
        "domains.DomainResult": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "auto_create_address": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_verification_attempt": {
                    "type": "string"
                },
                "next_verification_attempt": {
                    "type": "string"
                },
                "public_key": {
                    "type": "string"
                },
                "storage_enabled": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "verification_attempts": {
                    "type": "integer"
                },
                "verification_error": {
                    "type": "string"
                },
                "verification_status": {
                    "description": "Verification fields",
                    "type": "string"
                },
                "verification_token": {
                    "type": "string"
                },
                "webhook_config": {
                    "$ref": "#/definitions/domains.WebhookConfigResult"
                }
            }
        },
        "domains.UpdateDomainRequest": {
            "type": "object",
            "properties": {
                "auto_create_address": {
                    "type": "boolean"
                },
                "public_key": {
                    "type": "string"
                },
                "storage_enabled": {
                    "type": "boolean"
                },
                "verification_status": {
                    "$ref": "#/definitions/entities.VerificationStatus"
                }
            }
        },
        "domains.WebhookConfigRequest": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "secret": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "domains.WebhookConfigResult": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "secret": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "domains.planLimitExceededResponse": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "upgrade_url": {
                    "type": "string"
                }
            }
        },
...
        "entities.VerificationStatus": {
            "type": "string",
            "enum": [
                "pending",
                "validating",
                "verified",
                "failed",
                "expired"
            ],
            "x-enum-varnames": [
                "VerificationStatusPending",
                "VerificationStatusValidating",
                "VerificationStatusVerified",
                "VerificationStatusFailed",
                "VerificationStatusExpired"
            ]
        },
...
        "models.ErrorResponseBody": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
...
    }
}